amount change

+++ api/models.py
@@ -16,7 +16,7 @@ class Category(models.Model):
 class Product(models.Model):
     name =models.CharField(max_length=30, unique=True)
     slug = models.SlugField(null=True, blank=True)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.IntegerField()
     category = models.ForeignKey(Category, on_delete=models.CASCADE, related_name='category')
     image= models.ImageField(upload_to='photos/%Y/%m/%d', blank=True, null=True)
     thumbnail = models.ImageField(upload_to='photos/%Y/%m/%d', null=True, blank=True)
@@ -86,7 +86,7 @@ class Order(models.Model):
     place = models.CharField(max_length=255)
     phone = models.CharField(max_length=255)
     created_at = models.DateTimeField(auto_now_add=True)
-    total_amount=models.DecimalField(max_digits=9, decimal_places=2, blank=True, null=True)
+    total_amount=models.IntegerField(blank=True, null=True)
     
     def __str__(self):
         return self.orderId
